fix rotate

+++ definicje.cpp
@@ -17,7 +17,7 @@
 	};
 
 	bool drawOnlyStairs = true;
-	int rotate = 0;
+	float rotate = 0;
 	int const stairsCount = 78;
 
 	float movementSpeed = 0.1;
